fix: Quiet cmake warning for extracted timestamps
For zstd and hiredis downloaded from the Internet, set cmake policy
CMP0135 to NEW to set the timestamps of extracted archive members to the
time of extraction to suppress the cmake dev warning on newer versions
of cmake.

Check that the policy exists so that this works on older versions of
cmake.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>