fix: Let DirEntry::size_on_disk return likely size on disk
st_blocks in struct stat cannot be trusted for some filesystems that do
transparent compression or deduplication. For instance, ZFS can adjust
the block count some time in the future when compression is finished.

The only reasonable fix for this seems to be to just guess how much
space the file will take using st_size.

Fixes #1332.