feat: cmake -DOFFLINE=TRUE for offline builds
Add the cmake option OFFLINE, defaulting to the value of the standard
variable FETCHCONTENT_FULLY_DISCONNECTED, which is OFF by default, to
disable downloading anything from the internet.

When ON, set FETCHCONTENT_FULLY_DISCONNECTED to ON, ZSTD_FROM_INTERNET
to OFF and HIREDIS_FROM_INTERNET to OFF.

When downloading is OFF and either library is not found, throw an error
using find_package_handle_standard_args().

Add the option to INSTALL.md doc.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>