Add argument handling for Emscripten's --em-config
The arguments for the '--em-config' option for Emscripten were not
being processed correctly, causing ccache to incorrectly interpret the
following argument as an input file. This commit changes the argument
processing function to correctly handle '--em-config' along with the
required subsequent argument, improving compatibility with Emscripten's
command line interface.

Fixes #1457