fix: Fix binary patching of sysconfdir for Linux prebuilt binary
After fs::path-ification it no longer works to binary patch sysconfdir
at install time since the compiler apparently makes a note of the actual
length of SYSCONFDIR in

    const char k_sysconfdir[4096 + 1] = SYSCONFDIR;

and uses the length in the

    fs::path sysconfdir(k_sysconfdir);

call. When patching the SYSCONFDIR string the length of the original
string is used so that only the first strlen(SYSCONFDIR) part of the new
sysconfdir will be used.

Fix this by adding a pointer indirection which makes binary patching
work again.