fix: Ignore newer C++{20, 23} clang-tidy lints
This disables many of the newer modernization-related clang-tidy lints
that trigger when compiling as C++20 and C++23.

This fix is needed since the glaze dependency we will be using for C++20
modules support (for JSON parsing) compiles as C++23 which would cause
ccache compilation with -DENABLE_CLANG_TIDY=ON to fail because of the
newer lints.