Distinguish between Clang modules and C++20 modules
Clang features two forms of modules:

- Clang modules:
https://clang.llvm.org/docs/Modules.html

- C++20 modules:
https://clang.llvm.org/docs/StandardCPlusPlusModules.html

Clang modules pre-date C++20 modules and are based around the use of
"module.modulemap" files as a general mechanism for improving the
scalability of API modularization and imports across the C-family of
languages it supports, including C, C++, Objective-C, etc.

Due to complicated historical developments, the relationship between
these different forms of modules, including how their compilation models
differ, and exactly which form of modules is enabled with which
settings, can sometimes be unclear.

Prior to Clang 15, "-fmodules" will enable the use of Clang modules.
Whether or not Clang modules are actually used depends on the presence
and detection of "module.modulemap" files.

Since Clang 15, "-fmodules" will additionally enable C++20 standard
module support (which is also implicitly enabled with -std=c++20).

Note that the features are not mutually exclusive: Clang supports the
use of C++20 standard modules alongside the use of modulemaps.

Although ccache advertised support for modules, prior to this commit it
wasn't clear to the user which form of these modules was actually
supported by ccache and to what extent.

In this commit I've added documentation to clarify the situation and
have renamed some things that previously just referred to "modules" as
"clang_modules" instead.

This includes changing previous references to "C++ modules" also to
"Clang modules".

For further commits that develop support in ccache for C++20 standard
modules, which work with Clang, GCC, and MSVC, I adopt the naming scheme
of "C++20 modules" or cxx_modules where relevant.