Rewrite clang_modules test
This commit rewrites the clang_modules test to address several issues:

First, the way the test was written, it was not clear whether it was
supposed to be testing for C++ modules (as in the C++20 standard) or
Clang modules specifically.

In fact, it was only testing functionality related to Clang modules but
doing so with a file 'test1.cpp' compiled as C++, but which used
Objective-C++ module import syntax, i.e., "#import".

I suspect it was only accidental that this test even worked because
Clang ideally should have complained about invalid syntax in the C++
source file.

Indeed, Clang *did* complain if you tried compiling 'test1.cpp' on
Windows where '-fms-compatibility' is enabled by default, because then
it would treat the '#import' as non-C++ standard Microsoft MSVC syntax.

This problem Clang reported due to '-fms-compatibility' appears to be
the reason the $COMPILER_USES_MSVC check was added in de5778a. That
check has been removed since the test should run properly on windows.

Second, the Clang module mechanism which was being used here is just the
module.modulemap functionality. Clang supports modulemaps for all of the
C family languages it supports: C, C++, Objective-C, etc.

While the internal implementation of Clang's C++20 module support relies
on its more general Clang module support and related modulemap features,
the two are not directly related and can be used separately or together.

I've rewritten the test to use plain C with modulemaps in order to make
this distinction clear, especially as I intend to shortly add more tests
specific to C++20 standard modules which target all of the major
compilers with support for the feature.

I've also modified the check for support of Clang modules to scan the
--source-dependencies (-MD) output to look for module.modulemap, since
this is the crucial distinguishing characteristic as to whether or not
Clang modules are being used.