fix: Use weakly_canonical for checking ignore header paths
Previously, `remember_include_paths` used simpler string operations to
construct canonical paths for comparison.

This was seemingly sufficient for use with Clang and GCC on Linux but
unfortunately not sufficient for MSVC with some Windows paths.

In particular, the `test.direct` test would fail because of a prefixed
`.\\` in the path, which seems to be recently changed behavior because
until recently that test was not failing on CI.

Note that it would probably be better to canonicalize the ignored header
paths once when they are initially set in the context, but this would
require refactoring more code since canonicalization can fail (or we
could just ignore failure and do no transformation).