Add cmake option for enabling C++20 module features
This feature is gated behind a CMake option for the following reasons:

The glaze dependency we will use for fast JSON parsing requires a C++23
capable compiler. As of 2024, the three major compilers families
(Clang, GCC, MSVC) are supported.

The glaze dependency is currently only supported on little-endian
architectures. Big-endian support in glaze is currently being worked on
so this should be resolved eventually.