chore: Split util::split_once into util::split_once{,_to_views}
The util::split_once(std::string&&, char) version is a bit overly smart
in that it returns std::string instead of std::string_view for a
temporary std::string input. To reduce the risk for surprises, introduce
a util::split_once_into_views so that input lifetime is indicated in the
method name instead.