Refactor and unify compiler console processing
This commit combines multiple processing phases into a single phase:

- path rewriting for `base_dir`
- collecting includes from MSVC /showIncludes

The motivation for combining these is because for MSVC C++20 module
support, we also need to be able to process /sourceDependencies which
prints a JSON object to stdout.

Without this change, adding support for /sourceDependencies as a
separate phase would mean potentially processing stdout up to three
times per compilation.