fix: Do not rewrite -isysroot as relative path
Building relocatable pch files with Clang requires supplying the
`-isysroot` option which will turn include directories to be relative to
the build directory.

However, when combining this option with basedir, isysroot was turned
into a relative path which Clang does not accept in this case.

To fix this, we consider "relative" isysroot as cache key but invoke the
compiler with absolute paths.